Move builder to modules
