+++ themebuilder/cli/config.js
@@ -1,4 +1,5 @@
 var path = require("path");
+var readFile = require("./adapters/node-file-reader");
 
 var theme = process.env["npm_config_theme_name"] || "generic.light";
 var metadataFilePath = process.env["npm_config_metadata_file_path"] || "";
@@ -9,6 +10,8 @@ var themeName = theme.split(".")[0];
 var colorScheme = theme.split(".")[1] + (theme.split(".")[2] ? "-" + theme.split(".")[2] : "");
 var isBootstrap = false;
 var bootstrapVersion = 0;
+
+var metadataPromise = readFile(metadataFilePath);
 
 switch(path.extname(metadataFilePath)) {
     case ".scss":
@@ -25,7 +28,7 @@ var cssFileName = themeName + "." + colorScheme.replace(/-/, ".") + ".custom.css
 var cssFolderPath = "cli/artifacts/";
 
 module.exports = {
-    metadataFilePath: metadataFilePath,
+    metadataPromise: metadataPromise,
     fileFormat: fileFormat,
 
     themeName: themeName,
@@ -35,6 +38,9 @@ module.exports = {
 
     cssFileName: cssFileName,
     cssFolderPath: cssFolderPath,
-    swatchSelector: swatchSelector
+    swatchSelector: swatchSelector,
+
+    reader: readFile,
+    lessCompiler: require("less/lib/less-node")
 };
