+++ themebuilder/modules/builder.js
@@ -1,13 +1,10 @@
-var readFile = require("./adapters/node-file-reader");
-var lessCompiler = require("less/lib/less-node");
-
 var MetadataLoader = require("../modules/metadata-loader.js");
 var MetadataRepository = require("../modules/metadata-repository.js");
 var LessTemplateLoader = require("../modules/less-template-loader.js");
 var themes = require("../modules/themes.js");
 
 var processTheme = function(config, metadata, data) {
-    var lessTemplateLoader = new LessTemplateLoader(readFile, lessCompiler, config.swatchSelector);
+    var lessTemplateLoader = new LessTemplateLoader(config);
     if(config.isBootstrap) {
         var bootstrapMetadata = config.bootstrapVersion === 3 ?
             require("../data/bootstrap-metadata/bootstrap-metadata.js") :
@@ -35,10 +32,9 @@ var processTheme = function(config, metadata, data) {
 
 var buildTheme = function(config) {
     var metadataRepository = new MetadataRepository(new MetadataLoader());
-    var dataPromise = readFile(config.metadataFilePath);
     var repositoryPromise = metadataRepository.init(themes);
 
-    return Promise.all([dataPromise, repositoryPromise]).then(function(resolves) {
+    return Promise.all([config.metadataPromise, repositoryPromise]).then(function(resolves) {
         var data = resolves[0];
 
         var metadata = metadataRepository.getData({
